refactor(tests): split test_monai_spec.py into per-module files
